chore: コードの整理
